Fennel Proto REPL
Protocol-based REPL for Fennel Language.  This module embeds a
protocol that upgrades the REPL to work via structured messages.  New
client implementation utilizes the protocol for a more robust
experience and provides better API for extending it with new features.
Additionally, the support for Xref was added and implementation of
Eldoc was improved to be non-blocking.