remove eldoc support from fennel-repl
The old Eldoc implementation was inherently blocking, and could freeze
Emacs if the REPL is busy, and the cursor is hovering over something
that triggers the Eldoc.  Eldoc is now a feature of the
fennel-proto-repl module, old Eldoc-related variables are obsolete.
Additionally, this commit fixes prompt regexp escaping, and fixes
double REPL prompt by changing how the old-input obtained.