Prevent removing the undo history in the REPL buffer
Original code tried to preserve REPL prompt and its output after the
data was received from the process and printed but still provide
editing convenience to the user with correct undo history.  This
actually did not work properly, and undo history was discarded pretty
often, so the code was removed.  Also this commit sets the default
comment syntax.