Don't use dynamic variables for lexical-only purposes
compile-command is a defvar in compile.el which means it's a dynamic
var that can be overridden.  This is probably accidental, as the
variable is set and then immediately used, not expecting to modify
anything deeper in the call chain, as there are no other functions
that would later reference the compile-command var.  For some reason,
it breaks when calling the fennel-view-compilation function from
polymode's poly-org mode, hence the change.